Added linux makefiles. Added lib/getopt. Added lib/portaudio (Version 18.1). Added lib/portmixer. Applied flaac.patch to slimaudio_decoder_flac.c to allow 64-bit linux compiles. Applied sc7.patch to slimaudio_decoder.c to fix skipping issue with SC 7+. Removed automatic define of USE_SIGNALS_FOR_RESTART if not WIN32 in squeezeslave.c, must add define to appropriate Makefile. Modified version string to reflect export from slimdevices svn. Modified volume_control define from bool to use the enumerated type slimaudio_volume_t.  Fixes issue with VOLUME_DRIVER always being disabled regardless of command line option.

+++ src/slimaudio/slimaudio_decoder_flac.c
@@ -39,8 +39,7 @@
   #define DEBUGF(...)
   #define VDEBUGF(...)
 #endif
-
-static FLAC__StreamDecoderReadStatus flac_read_callback(const FLAC__StreamDecoder *decoder, FLAC__byte buffer[], unsigned *bytes, void *client_data);
+static FLAC__StreamDecoderReadStatus flac_read_callback(const FLAC__StreamDecoder *decoder, FLAC__byte buffer[], size_t *bytes, void *client_data);
 static FLAC__StreamDecoderWriteStatus flac_write_callback(const FLAC__StreamDecoder *decoder, const FLAC__Frame *frame, const FLAC__int32 * const buffer[], void *client_data);
 static void flac_metadata_callback(const FLAC__StreamDecoder *decoder, const FLAC__StreamMetadata *metadata, void *client_data);
 static void flac_error_callback(const FLAC__StreamDecoder *decoder, FLAC__StreamDecoderErrorStatus status, void *client_data);
@@ -98,8 +97,7 @@ int slimaudio_decoder_flac_process(slimaudio_t *audio) {
 	return (b == true ? 0 : -1);
 }
 
-
-static FLAC__StreamDecoderReadStatus flac_read_callback(const FLAC__StreamDecoder *decoder, FLAC__byte buffer[], unsigned *bytes, void *client_data) {
+static FLAC__StreamDecoderReadStatus flac_read_callback(const FLAC__StreamDecoder *decoder, FLAC__byte buffer[], size_t *bytes, void *client_data) {
 	slimaudio_t *audio = (slimaudio_t *) client_data;
 	
 	pthread_mutex_lock(&audio->decoder_mutex);
